backend/modules/integrations/oauth.py
"""
OAuth Integration Module

TODO: Implement OAuth flow for Google services using Composio
Contributors: Handle connection management
"""

from typing import Dict, Any, Optional, List
import os
from core.config import settings


class OAuthManager:
    """
    OAuth Manager using Composio
    
    TODO: Implement OAuth connection management
    """
    
    def __init__(self):
        """
        Initialize OAuth Manager
        
        TODO: Setup Composio client
        - Load API key from settings
        - Initialize Composio SDK
        """
        self.composio_api_key = settings.COMPOSIO_API_KEY
        self.auth_configs = {
            "gmail": settings.AC_GMAIL,
            "calendar": settings.AC_GOOGLE_CALENDAR,
            "sheets": settings.AC_GOOGLE_SHEETS
        }
    
    
    def get_composio_client(self):
        """
        Get Composio client instance
        
        TODO: Implement client creation
        - Import from composio
        - Return configured client
        
        Returns:
            Composio client
        """
        from composio import Composio
        return Composio(api_key=self.composio_api_key)
    
    
    def initiate_connection(
        self,
        service: str,
        user_id: str,
        return_url: Optional[str] = None
    ) -> Dict[str, Any]:
        """
        Initiate OAuth connection for a service
        
        TODO: Implement OAuth initiation
        - Use Composio connected_accounts.initiate()
        - Get auth_config_id for service
        - Return redirect URL
        
        Reference: Your previous working code
        
        Args:
            service: Service name (gmail, calendar, sheets)
            user_id: User ID for Composio entity
            return_url: URL to return to after OAuth
            
        Returns:
            OAuth details with redirect URL
        """
        if service not in self.auth_configs:
            raise ValueError(f"Unknown service: {service}")
        
        auth_config_id = self.auth_configs[service]
        if not auth_config_id:
            raise ValueError(f"Auth config not set for {service}")
        
        # TODO: Implement using Composio
        # composio = self.get_composio_client()
        # connection_request = composio.connected_accounts.initiate(
        #     user_id=user_id,
        #     auth_config_id=auth_config_id,
        #     allow_multiple=True
        # )
        # return {
        #     "redirect_url": connection_request.redirectUrl,
        #     "status": "initiated"
        # }
        
        raise NotImplementedError("Implement OAuth initiation using Composio")
    
    
    def handle_callback(
        self,
        code: str,
        state: str
    ) -> Dict[str, Any]:
        """
        Handle OAuth callback
        
        TODO: Implement callback handling
        - Composio handles this automatically
        - Just verify connection was created
        - Return connection status
        
        Note: Composio manages the callback internally
        
        Args:
            code: OAuth code
            state: State parameter
            
        Returns:
            Connection status
        """
        # Composio handles callback automatically
        # This endpoint can just redirect to success page
        return {"status": "connected"}
    
    
    def get_connection_status(
        self,
        user_id: str,
        service: Optional[str] = None
    ) -> Dict[str, Any]:
        """
        Get connection status for user
        
        TODO: Implement status check
        - Use Composio connected_accounts.list()
        - Filter by user_id
        - Return status for each service
        
        Args:
            user_id: User ID
            service: Optional specific service
            
        Returns:
            Connection status
        """
        # TODO: Implement using Composio
        # composio = self.get_composio_client()
        # connections = composio.connected_accounts.list(entity_ids=[user_id])
        # ... process and return
        
        raise NotImplementedError("Implement connection status check using Composio")
    
    
    def disconnect_service(
        self,
        user_id: str,
        service: str
    ) -> Dict[str, Any]:
        """
        Disconnect a service
        
        TODO: Implement disconnection
        - Find connection for user and service
        - Delete connection
        - Revoke OAuth tokens
        
        Args:
            user_id: User ID
            service: Service to disconnect
            
        Returns:
            Disconnection status
        """
        raise NotImplementedError("Implement service disconnection using Composio")
    
    
    def get_user_connections(self, user_id: str) -> List[Dict[str, Any]]:
        """
        Get all connections for a user
        
        TODO: Implement connection listing
        - List all connected services
        - Return service names and status
        
        Args:
            user_id: User ID
            
        Returns:
            List of connections
        """
        raise NotImplementedError("Implement connection listing using Composio")
