backend/requirements.txt
# FastAPI and Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# AI and LLM
openai==1.3.0
composio-core==0.6.0

# Document Processing
PyPDF2==3.0.1
pdfplumber==0.10.3
python-docx==1.1.0
pytesseract==0.3.10

# NLP and ML
spacy==3.7.2
sentence-transformers==2.2.2
scikit-learn==1.3.2

# Vector Database (choose one)
chromadb==0.4.18
# pinecone-client==2.2.4
# weaviate-client==3.25.3

# Database (if needed)
# sqlalchemy==2.0.23
# alembic==1.12.1

# Utilities
requests==2.31.0
aiohttp==3.9.1
pandas==2.1.3
numpy==1.26.2

# Testing (optional)
# pytest==7.4.3
# pytest-asyncio==0.21.1
# httpx==0.25.2
